refactor ipfix modules: create ie.rs and flow_record.rs
